changes to login process
